cleaned up old module implementations
