--- fpgaconvnet/models/modules/ThresholdedReLU.py
@@ -1,56 +1,0 @@
-"""
-.. figure:: ../../../figures/relu_diagram.png
-"""
-
-import math
-import os
-from dataclasses import dataclass, field
-
-import numpy as np
-import pydot
-
-from fpgaconvnet.models.modules import Module, MODULE_FONTSIZE
-
-@dataclass
-class ThresholdedReLU(Module):
-
-    def __post_init__(self):
-        pass
-
-    def rsc(self, coef=None, model=None):
-        """
-        Returns
-        -------
-        dict
-            estimated resource usage of the module. Uses the
-            resource coefficients for the estimate.
-        """
-        return {
-            "LUT"   : 16,
-            "FF"    : 35,
-            "BRAM"  : 0,
-            "DSP"   : 0
-        }
-
-    def visualise(self, name):
-        return pydot.Node(name,label="relu", shape="box",
-                style="filled", fillcolor="dimgrey",
-                fontsize=MODULE_FONTSIZE)
-
-    def functional_model(self, data):
-        # check input dimensionality
-        assert data.shape[0] == self.rows    , "ERROR: invalid row dimension"
-        assert data.shape[1] == self.cols    , "ERROR: invalid column dimension"
-        assert data.shape[2] == self.channels, "ERROR: invalid channel dimension"
-
-        out = np.ndarray((
-            self.rows,
-            self.cols,
-            self.channels),dtype=float)
-
-        for index,_ in np.ndenumerate(out):
-            out[index] = max(data[index],self.threshold)
-
-        return out
-
-
