+++ fpgaconvnet/models/modules/global_pool/chisel.py
@@ -68,22 +68,40 @@ class GlobalPoolChisel(ModuleChiselBase):
     def resource_parameters_heuristics(self) -> dict[str, list[int]]:
         return {
             "Logic_LUT" : np.array([
-                    self.data_t.width,
-                ]),
+                self.acc_width, # adder
+                self.data_width, # adder
+                int2bits(self.channels), # channel_cntr
+                int2bits(self.rows*self.cols), # spatial cntr
+                1,
+            ]),
             "LUT_RAM"   : np.array([
-                    self.data_t.width*self.channels,
-                ]),
+                # queue_lutram_resource_model(
+                #     4, self.data_width), # buffer
+            ]),
             "LUT_SR"    : np.array([0]),
             "FF"        : np.array([
-                    self.data_t.width,
-                    int2bits(self.channels),
-                    1,
-                ]),
+                self.data_width, # input cache
+                int2bits(self.channels), # channel_cntr
+                int2bits(self.rows*self.cols), # spatial cntr
+                self.acc_width*self.channels, # accumulation reg
+                1, # other registers
+            ]),
             "DSP"       : np.array([0]),
             "BRAM36"    : np.array([0]),
             "BRAM18"    : np.array([0]),
         }
 
+    # def rsc(self, coef=None, model=None):
+
+    #     # get the regression model estimation
+    #     rsc = Module.rsc(self, coef, model)
+
+    #     if self.regression_model == "linear_regression":
+    #         # get the dsp usage
+    #         rsc["DSP"] = dsp_multiplier_resource_model(
+    #                 self.data_width, self.acc_width)
+
+    #     return rsc
 
     def functional_model(self, data):
