+++ fpgaconvnet/models/modules/glue/hls.py
@@ -75,14 +75,6 @@ class GlueHLS(GlueHLSBase):
         return [ self.rows, self.cols, self.filters, self.coarse_in,
                 self.coarse_out, self.coarse_group, self.data_t.width, self.acc_t.width ]
 
-    def resource_parameters_heuristics(self) -> dict[str, list[int]]:
-        return {
-            "LUT"   : np.array([1]),
-            "FF"    : np.array([1]),
-            "DSP"   : np.array([0]),
-            "BRAM"  : np.array([0]),
-        }
-
 @dataclass
 class GlueHLS3D(ModuleHLS3DBase, GlueHLSBase):
 
@@ -99,12 +91,4 @@ class GlueHLS3D(ModuleHLS3DBase, GlueHLSBase):
     def resource_parameters(self) -> list[int]:
         return [ self.rows, self.cols, self.depth, self.filters, self.coarse_in,
                 self.coarse_out, self.coarse_group, self.data_t.width, self.acc_t.width ]
-
-    def resource_parameters_heuristics(self) -> dict[str, list[int]]:
-        return {
-            "LUT"   : np.array([1]),
-            "FF"    : np.array([1]),
-            "DSP"   : np.array([0]),
-            "BRAM"  : np.array([0]),
-        }
